requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
aio_pika